[Security Solution] enabling all expanded flyout Cypress tests
