+++ x-pack/plugins/security_solution/cypress/e2e/investigations/alerts/expandable_flyout/alert_details_url_sync.cy.ts
@@ -8,50 +8,48 @@
 import { getNewRule } from '../../../../objects/rule';
 import { cleanKibana } from '../../../../tasks/common';
 import { waitForAlertsToPopulate } from '../../../../tasks/create_new_rule';
-import { expandFirstAlertExpandableFlyout } from '../../../../tasks/document_expandable_flyout';
 import { login, visit } from '../../../../tasks/login';
 import { createRule } from '../../../../tasks/api_calls/rules';
 import { ALERTS_URL } from '../../../../urls/navigation';
-import {
-  DOCUMENT_DETAILS_FLYOUT_CLOSE_BUTTON,
-  DOCUMENT_DETAILS_FLYOUT_HEADER_TITLE,
-} from '../../../../screens/document_expandable_flyout';
+import { closeFlyout } from '../../../../tasks/expandable_flyout/alert_details_right_panel';
+import { expandFirstAlertExpandableFlyout } from '../../../../tasks/expandable_flyout/common';
+import { DOCUMENT_DETAILS_FLYOUT_HEADER_TITLE } from '../../../../screens/expandable_flyout/alert_details_right_panel';
 
-// Skipping these for now as the feature is protected behind a feature flag set to false by default
-// To run the tests locally, add 'securityFlyoutEnabled' in the Cypress config.ts here https://github.com/elastic/kibana/blob/main/x-pack/test/security_solution_cypress/config.ts#L50
-describe.skip(
+describe(
   'Expandable flyout state sync',
   { env: { ftrConfig: { enableExperimental: ['securityFlyoutEnabled'] } } },
   () => {
     const rule = getNewRule();
 
-    before(() => {
+    beforeEach(() => {
       cleanKibana();
       login();
       createRule(rule);
       visit(ALERTS_URL);
       waitForAlertsToPopulate();
+    });
+
+    it('should test flyout url sync', () => {
+      cy.url().should('not.include', 'eventFlyout');
+
       expandFirstAlertExpandableFlyout();
-    });
 
-    it('should serialize its state to url', () => {
-      cy.url().should('include', 'eventFlyout');
-      cy.get(DOCUMENT_DETAILS_FLYOUT_HEADER_TITLE).should('be.visible').and('have.text', rule.name);
-    });
-
-    it('should reopen the flyout after browser refresh', () => {
-      cy.reload();
+      cy.log('should serialize its state to url');
 
       cy.url().should('include', 'eventFlyout');
       cy.get(DOCUMENT_DETAILS_FLYOUT_HEADER_TITLE).should('be.visible').and('have.text', rule.name);
-    });
 
-    it('should clear the url state when flyout is closed', () => {
+      cy.log('should reopen the flyout after browser refresh');
+
       cy.reload();
+      waitForAlertsToPopulate();
 
+      cy.url().should('include', 'eventFlyout');
       cy.get(DOCUMENT_DETAILS_FLYOUT_HEADER_TITLE).should('be.visible').and('have.text', rule.name);
 
-      cy.get(DOCUMENT_DETAILS_FLYOUT_CLOSE_BUTTON).click();
+      cy.log('should clear the url state when flyout is closed');
+
+      closeFlyout();
 
       cy.url().should('not.include', 'eventFlyout');
     });
