+++ x-pack/plugins/security_solution/cypress/screens/expandable_flyout/alert_details_right_panel_json_tab.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { getDataTestSubjectSelector } from '../../helpers/common';
+import { JSON_TAB_CONTENT_TEST_ID } from '../../../public/flyout/right/tabs/test_ids';
+
+export const DOCUMENT_DETAILS_FLYOUT_JSON_TAB_CONTENT =
+  getDataTestSubjectSelector(JSON_TAB_CONTENT_TEST_ID);
