+++ x-pack/plugins/security_solution/cypress/screens/expandable_flyout/alert_details_right_panel_overview_tab.ts
@@ -1,0 +1,131 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { getDataTestSubjectSelector } from '../../helpers/common';
+import {
+  ANALYZER_TREE_TEST_ID,
+  DESCRIPTION_DETAILS_TEST_ID,
+  DESCRIPTION_EXPAND_BUTTON_TEST_ID,
+  DESCRIPTION_SECTION_CONTENT_TEST_ID,
+  DESCRIPTION_SECTION_HEADER_TEST_ID,
+  DESCRIPTION_TITLE_TEST_ID,
+  ENTITIES_CONTENT_TEST_ID,
+  ENTITIES_HEADER_TEST_ID,
+  ENTITIES_VIEW_ALL_BUTTON_TEST_ID,
+  ENTITY_PANEL_CONTENT_TEST_ID,
+  ENTITY_PANEL_HEADER_TEST_ID,
+  HIGHLIGHTED_FIELDS_DETAILS_TEST_ID,
+  HIGHLIGHTED_FIELDS_GO_TO_TABLE_LINK,
+  HIGHLIGHTED_FIELDS_TITLE_TEST_ID,
+  INSIGHTS_CORRELATIONS_CONTENT_TEST_ID,
+  INSIGHTS_CORRELATIONS_TITLE_TEST_ID,
+  INSIGHTS_CORRELATIONS_VALUE_TEST_ID,
+  INSIGHTS_CORRELATIONS_VIEW_ALL_BUTTON_TEST_ID,
+  INSIGHTS_HEADER_TEST_ID,
+  INSIGHTS_PREVALENCE_CONTENT_TEST_ID,
+  INSIGHTS_PREVALENCE_TITLE_TEST_ID,
+  INSIGHTS_PREVALENCE_VALUE_TEST_ID,
+  INSIGHTS_PREVALENCE_VIEW_ALL_BUTTON_TEST_ID,
+  INSIGHTS_THREAT_INTELLIGENCE_CONTENT_TEST_ID,
+  INSIGHTS_THREAT_INTELLIGENCE_TITLE_TEST_ID,
+  INSIGHTS_THREAT_INTELLIGENCE_VALUE_TEST_ID,
+  INSIGHTS_THREAT_INTELLIGENCE_VIEW_ALL_BUTTON_TEST_ID,
+  INVESTIGATION_GUIDE_BUTTON_TEST_ID,
+  INVESTIGATION_SECTION_CONTENT_TEST_ID,
+  INVESTIGATION_SECTION_HEADER_TEST_ID,
+  MITRE_ATTACK_DETAILS_TEST_ID,
+  MITRE_ATTACK_TITLE_TEST_ID,
+  REASON_DETAILS_TEST_ID,
+  REASON_TITLE_TEST_ID,
+  VISUALIZATIONS_SECTION_HEADER_TEST_ID,
+} from '../../../public/flyout/right/components/test_ids';
+
+/* Description section */
+
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_DESCRIPTION_SECTION_HEADER =
+  getDataTestSubjectSelector(DESCRIPTION_SECTION_HEADER_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_DESCRIPTION_SECTION_CONTENT =
+  getDataTestSubjectSelector(DESCRIPTION_SECTION_CONTENT_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_DESCRIPTION_TITLE =
+  getDataTestSubjectSelector(DESCRIPTION_TITLE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_DESCRIPTION_DETAILS = getDataTestSubjectSelector(
+  DESCRIPTION_DETAILS_TEST_ID
+);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_DESCRIPTION_EXPAND_BUTTON =
+  getDataTestSubjectSelector(DESCRIPTION_EXPAND_BUTTON_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_REASON_TITLE =
+  getDataTestSubjectSelector(REASON_TITLE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_REASON_DETAILS =
+  getDataTestSubjectSelector(REASON_DETAILS_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_MITRE_ATTACK_TITLE = getDataTestSubjectSelector(
+  MITRE_ATTACK_TITLE_TEST_ID
+);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_MITRE_ATTACK_DETAILS = getDataTestSubjectSelector(
+  MITRE_ATTACK_DETAILS_TEST_ID
+);
+
+/* Investigation section */
+
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INVESTIGATION_SECTION_HEADER =
+  getDataTestSubjectSelector(INVESTIGATION_SECTION_HEADER_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INVESTIGATION_SECTION_CONTENT =
+  getDataTestSubjectSelector(INVESTIGATION_SECTION_CONTENT_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_HIGHLIGHTED_FIELDS_HEADER_TITLE =
+  getDataTestSubjectSelector(HIGHLIGHTED_FIELDS_TITLE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_HIGHLIGHTED_FIELDS_DETAILS =
+  getDataTestSubjectSelector(HIGHLIGHTED_FIELDS_DETAILS_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_HIGHLIGHTED_FIELDS_GO_TO_TABLE_LINK =
+  getDataTestSubjectSelector(HIGHLIGHTED_FIELDS_GO_TO_TABLE_LINK);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INVESTIGATION_GUIDE_BUTTON =
+  getDataTestSubjectSelector(INVESTIGATION_GUIDE_BUTTON_TEST_ID);
+
+/* Insights section */
+
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_SECTION_HEADER =
+  getDataTestSubjectSelector(INSIGHTS_HEADER_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_ENTITIES_HEADER =
+  getDataTestSubjectSelector(ENTITIES_HEADER_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_ENTITIES_CONTENT =
+  getDataTestSubjectSelector(ENTITIES_CONTENT_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_VIEW_ALL_ENTITIES_BUTTON =
+  getDataTestSubjectSelector(ENTITIES_VIEW_ALL_BUTTON_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_ENTITY_PANEL_HEADER =
+  getDataTestSubjectSelector(ENTITY_PANEL_HEADER_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_ENTITY_PANEL_CONTENT =
+  getDataTestSubjectSelector(ENTITY_PANEL_CONTENT_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_THREAT_INTELLIGENCE_HEADER =
+  getDataTestSubjectSelector(INSIGHTS_THREAT_INTELLIGENCE_TITLE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_THREAT_INTELLIGENCE_CONTENT =
+  getDataTestSubjectSelector(INSIGHTS_THREAT_INTELLIGENCE_CONTENT_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_THREAT_INTELLIGENCE_VALUES =
+  getDataTestSubjectSelector(INSIGHTS_THREAT_INTELLIGENCE_VALUE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_THREAT_INTELLIGENCE_VIEW_ALL_BUTTON =
+  getDataTestSubjectSelector(INSIGHTS_THREAT_INTELLIGENCE_VIEW_ALL_BUTTON_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_CORRELATIONS_HEADER =
+  getDataTestSubjectSelector(INSIGHTS_CORRELATIONS_TITLE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_CORRELATIONS_CONTENT =
+  getDataTestSubjectSelector(INSIGHTS_CORRELATIONS_CONTENT_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_CORRELATIONS_VALUES =
+  getDataTestSubjectSelector(INSIGHTS_CORRELATIONS_VALUE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_CORRELATIONS_VIEW_ALL_BUTTON =
+  getDataTestSubjectSelector(INSIGHTS_CORRELATIONS_VIEW_ALL_BUTTON_TEST_ID);
+
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_PREVALENCE_HEADER =
+  getDataTestSubjectSelector(INSIGHTS_PREVALENCE_TITLE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_PREVALENCE_CONTENT =
+  getDataTestSubjectSelector(INSIGHTS_PREVALENCE_CONTENT_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_PREVALENCE_VALUES =
+  getDataTestSubjectSelector(INSIGHTS_PREVALENCE_VALUE_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_INSIGHTS_PREVALENCE_VIEW_ALL_BUTTON =
+  getDataTestSubjectSelector(INSIGHTS_PREVALENCE_VIEW_ALL_BUTTON_TEST_ID);
+
+/* Visualization section */
+
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_VISUALIZATIONS_SECTION_HEADER =
+  getDataTestSubjectSelector(VISUALIZATIONS_SECTION_HEADER_TEST_ID);
+export const DOCUMENT_DETAILS_FLYOUT_OVERVIEW_TAB_ANALYZER_TREE =
+  getDataTestSubjectSelector(ANALYZER_TREE_TEST_ID);
