+++ x-pack/plugins/security_solution/public/flyout/right/components/correlations_overview.tsx
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import React, { useCallback } from 'react';
+import React, { useCallback, useMemo } from 'react';
 import { EuiButtonEmpty, EuiFlexGroup, EuiPanel } from '@elastic/eui';
 import { useExpandableFlyoutContext } from '@kbn/expandable-flyout';
 import { InsightsSummaryRow } from './insights_summary_row';
@@ -44,6 +44,20 @@ export const CorrelationsOverview: React.FC = () => {
     scopeId,
   });
 
+  const correlationRows = useMemo(
+    () =>
+      data.map((d) => (
+        <InsightsSummaryRow
+          icon={d.icon}
+          value={d.value}
+          text={d.text}
+          data-test-subj={INSIGHTS_CORRELATIONS_TEST_ID}
+          key={`correlation-row-${d.text}`}
+        />
+      )),
+    [data]
+  );
+
   return (
     <InsightsSubSection
       loading={loading}
@@ -53,14 +67,7 @@ export const CorrelationsOverview: React.FC = () => {
     >
       <EuiPanel hasShadow={false} hasBorder={true} paddingSize="s">
         <EuiFlexGroup direction="column" gutterSize="none">
-          {data.map((d) => (
-            <InsightsSummaryRow
-              icon={d.icon}
-              value={d.value}
-              text={d.text}
-              data-test-subj={INSIGHTS_CORRELATIONS_TEST_ID}
-            />
-          ))}
+          {correlationRows}
         </EuiFlexGroup>
       </EuiPanel>
       <EuiButtonEmpty
