+++ x-pack/plugins/security_solution/public/flyout/right/hooks/use_fetch_related_alerts_by_ancestry.ts
@@ -7,6 +7,7 @@
 
 import type { TimelineEventsDetailsItem } from '@kbn/timelines-plugin/common';
 import { find } from 'lodash/fp';
+import { useMemo } from 'react';
 import { useAlertPrevalenceFromProcessTree } from '../../../common/containers/alerts/use_alert_prevalence_from_process_tree';
 import { isActiveTimeline } from '../../../helpers';
 
@@ -47,23 +48,30 @@ export const useFetchRelatedAlertsByAncestry = ({
   dataFormattedForFieldBrowser,
   scopeId,
 }: UseFetchRelatedAlertsByAncestryParams): UseFetchRelatedAlertsByAncestryResult => {
-  const originalDocumentId = find(
-    { category: 'kibana', field: 'kibana.alert.ancestors.id' },
-    dataFormattedForFieldBrowser
+  const documentId = useMemo(() => {
+    const originalDocumentId = find(
+      { category: 'kibana', field: 'kibana.alert.ancestors.id' },
+      dataFormattedForFieldBrowser
+    );
+    const { values } = originalDocumentId ?? { values: [] };
+    return Array.isArray(values) ? values[0] : '';
+  }, [dataFormattedForFieldBrowser]);
+
+  const { values: indices } = useMemo(
+    () =>
+      find(
+        { category: 'kibana', field: 'kibana.alert.rule.parameters.index' },
+        dataFormattedForFieldBrowser
+      ) || { values: [] },
+    [dataFormattedForFieldBrowser]
   );
-  const originalDocumentIndex = find(
-    { category: 'kibana', field: 'kibana.alert.rule.parameters.index' },
-    dataFormattedForFieldBrowser
-  );
+
   const isActiveTimelines = isActiveTimeline(scopeId ?? '');
 
-  const { values: indices } = originalDocumentIndex || { values: [] };
-  const { values: wrappedDocumentId } = originalDocumentId || { values: [] };
-  const documentId = Array.isArray(wrappedDocumentId) ? wrappedDocumentId[0] : '';
   const { loading, error, alertIds } = useAlertPrevalenceFromProcessTree({
     isActiveTimeline: isActiveTimelines,
     documentId,
-    indices: indices ?? [],
+    indices: indices || [],
   });
 
   return {
