+++ x-pack/plugins/security_solution/public/flyout/right/hooks/use_fetch_related_alerts_by_same_source_event.ts
@@ -7,6 +7,7 @@
 
 import type { TimelineEventsDetailsItem } from '@kbn/timelines-plugin/common';
 import { find } from 'lodash/fp';
+import { useMemo } from 'react';
 import { useAlertPrevalence } from '../../../common/containers/alerts/use_alert_prevalence';
 import { isActiveTimeline } from '../../../helpers';
 
@@ -46,11 +47,15 @@ export const useFetchRelatedAlertsBySameSourceEvent = ({
   dataFormattedForFieldBrowser,
   scopeId,
 }: UseFetchRelatedAlertsBySameSourceEventParams): UseFetchRelatedAlertsBySameSourceEventResult => {
-  const sourceEventField = find(
-    { category: 'kibana', field: 'kibana.alert.original_event.id' },
-    dataFormattedForFieldBrowser
+  const { field, values } = useMemo(
+    () =>
+      find(
+        { category: 'kibana', field: 'kibana.alert.original_event.id' },
+        dataFormattedForFieldBrowser
+      ) || { field: '', values: [] },
+    [dataFormattedForFieldBrowser]
   );
-  const { field, values } = sourceEventField || { field: '', values: [] };
+
   const { loading, error, count, alertIds } = useAlertPrevalence({
     field,
     value: values,
